+++ crates/forge/src/main.rs
@@ -18,7 +18,6 @@ use scarb_api::{
 };
 use scarb_metadata::{Metadata, MetadataCommand, PackageMetadata};
 use scarb_ui::args::PackagesFilter;
-use universal_sierra_compiler_api::UniversalSierraCompilerCommand;
 
 use forge::block_number_map::BlockNumberMap;
 use semver::{Comparator, Op, Version, VersionReq};
@@ -274,7 +273,6 @@ fn main_execution() -> Result<bool> {
     let cli = Cli::parse();
 
     ScarbCommand::new().ensure_available()?;
-    UniversalSierraCompilerCommand::ensure_available()?;
 
     match cli.subcommand {
         ForgeSubcommand::Init { name } => {