+++ crates/sncast/tests/helpers/fixtures.rs
@@ -1,12 +1,14 @@
-use crate::helpers::constants::{ACCOUNT_FILE_PATH, CONTRACTS_DIR, DEVNET_ENV_FILE, URL};
+use crate::helpers::constants::{
+    ACCOUNT_FILE_PATH, CONTRACTS_DIR, DEVNET_ENV_FILE, DEVNET_OZ_CLASS_HASH, URL,
+};
 use camino::Utf8PathBuf;
 use primitive_types::U256;
 use serde::de::DeserializeOwned;
 use serde::Deserialize;
 use serde_json::{json, Map, Value};
-use sncast::{apply_optional, get_keystore_password};
+use sncast::{apply_optional, get_chain_id, get_keystore_password};
 use sncast::{get_account, get_provider, parse_number};
-use starknet::accounts::{Account, Call, Execution};
+use starknet::accounts::{Account, AccountFactory, Call, Execution, OpenZeppelinAccountFactory};
 use starknet::contract::ContractFactory;
 use starknet::core::types::contract::{CompiledClass, SierraClass};
 use starknet::core::types::TransactionReceipt;
@@ -15,7 +17,7 @@ use starknet::core::utils::get_contract_address;
 use starknet::core::utils::get_selector_from_name;
 use starknet::providers::jsonrpc::HttpTransport;
 use starknet::providers::JsonRpcClient;
-use starknet::signers::SigningKey;
+use starknet::signers::{LocalWallet, SigningKey};
 use std::env;
 use std::fs;
 use std::fs::OpenOptions;
@@ -67,6 +69,46 @@ pub async fn declare_contract(account: &str, path: &str, shortname: &str) -> Fie
     write_devnet_env(format!("{shortname}_CLASS_HASH").as_str(), &class_hash);
     write_devnet_env(format!("{shortname}_DECLARE_HASH").as_str(), &tx_hash);
     class_hash
+}
+
+pub async fn deploy_keystore_account() {
+    let keystore_path = "tests/data/keystore/predeployed_key.json";
+    let account_path = "tests/data/keystore/predeployed_account.json";
+    let private_key =
+        SigningKey::from_keystore(keystore_path, "123").expect("Could not get the private_key");
+
+    let provider = get_provider(URL).expect("Could not get the provider");
+    let chain_id = get_chain_id(&provider)
+        .await
+        .expect("Could not get chain_id from provider");
+
+    let contents =
+        std::fs::read_to_string(account_path).expect("Failed to read keystore account file");
+    let items: serde_json::Value = serde_json::from_str(&contents)
+        .unwrap_or_else(|_| panic!("Failed to parse keystore account file at = {account_path}"));
+
+    let factory = OpenZeppelinAccountFactory::new(
+        parse_number(DEVNET_OZ_CLASS_HASH).expect("Could not parse DEVNET_OZ_CLASS_HASH"),
+        chain_id,
+        LocalWallet::from_signing_key(private_key),
+        provider,
+    )
+    .await
+    .expect("Could not create Account Factory");
+
+    mint_token(
+        items["deployment"]["address"]
+            .as_str()
+            .expect("Could not get address"),
+        9_999_999_999_999_999_999,
+    )
+    .await;
+
+    factory
+        .deploy(parse_number("0xa5d90c65b1b1339").expect("Could not parse salt"))
+        .send()
+        .await
+        .expect("Could not deploy keystore account");
 }
 
 pub async fn declare_deploy_contract(account: &str, path: &str, shortname: &str) {