+++ CHANGELOG.md
@@ -13,7 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `--detailed-resources` flag for displaying additional info about used resources
 - `store` and `load` cheatcodes
-- dependency on [universal-sierra-compiler](https://github.com/software-mansion/universal-sierra-compiler/blob/master/scripts/install.sh) binary
+
+#### Fixed
+
+- Error message for tests that should panic but pass
 
 ## [0.14.0] - 2024-01-11